+++ source/Private/Invoke-LaunchTreeUnelevatedProcess.ps1
@@ -1,0 +1,28 @@
+function Invoke-LaunchTreeUnelevatedProcess {
+    [CmdletBinding()]
+    [OutputType([int])]
+    param(
+        [Parameter(Mandatory)]
+        [ValidateNotNullOrEmpty()]
+        [string] $ApplicationPath,
+
+        [Parameter(Mandatory)]
+        [AllowEmptyString()]
+        [string] $Arguments,
+
+        [Parameter(Mandatory)]
+        [ValidateNotNullOrEmpty()]
+        [string] $WorkingDirectory,
+
+        [Parameter(Mandatory)]
+        [int] $TimeoutMilliseconds
+    )
+
+    Initialize-LaunchTreeUnelevatedProcess
+    [LaunchTree.UnelevatedProcess]::Run(
+        $ApplicationPath,
+        $Arguments,
+        $WorkingDirectory,
+        $TimeoutMilliseconds
+    )
+}